requirements.txt
aiogram==2.25.1
google-generativeai 